pyproject.toml
[tool.poetry]
name = "fast-api-project"
version = "0.1.0"
description = ""
authors = ["P-Roq"]
readme = "README.md"
packages = [{include = "fast_api_project"}]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "^0.103.2"}
uvicorn = "^0.23.2"
sqlalchemy = "1.3.18"
pandas = "2.0.0"
python-dotenv = "^1.0.0"
bandit = "^1.7.5"
mysql-connector-python = "^8.1.0"
mysql-connector-python-rf = "^2.2.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
alembic = "^1.12.1"
pytest = "^7.4.3"
pymysql = "^1.1.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
